pyproject.toml
[project]
name = "compress-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0",
    "pillow>=10.0",
    "ffmpeg-python>=0.2",
    "celery>=5.3",
    "sqlalchemy>=2.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
]
